pyproject.toml
[project]
name = "pdx-traffic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs>=21.3.0",
    "dagster>=1.9.6",
    "duckdb>=1.1.3",
    "httpx>=0.20.0,<0.28.0",
    "lonboard>=0.10.3",
    "polars>=1.17.1",
    "python-dateutil>=2.8.0,<3.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "trip-check-api-v-1-client",
]

[tool.uv.sources]
trip-check-api-v-1-client = { path = "trip-check-api-v-1-client", editable = true }

[dependency-groups]
dev = [
    "dagster-webserver>=1.9.6",
]


[tool.dagster]
module_name = "dags"